+++ src/main/java/roomescape/theme/service/ThemeService.java
@@ -2,8 +2,9 @@ package roomescape.theme.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import roomescape.global.exception.ConflictException;
+import roomescape.reservation.repository.ReservationRepository;
 import roomescape.theme.domain.Theme;
-import roomescape.global.exception.InvalidRequestException;
 import roomescape.theme.repository.ThemeRepository;
 
 import java.time.LocalDate;
@@ -12,32 +13,30 @@ import java.util.List;
 @Service
 public class ThemeService {
     private final ThemeRepository themeRepository;
+    private final ReservationRepository reservationRepository;
 
-    public ThemeService(ThemeRepository themeRepository) {
+    public ThemeService(ThemeRepository themeRepository, ReservationRepository reservationRepository) {
         this.themeRepository = themeRepository;
+        this.reservationRepository = reservationRepository;
     }
 
     @Transactional
     public Theme create(String name, String description, String thumbnail) {
-        validateNotDuplicated(name);
+        if (themeRepository.existsByName(name)) {
+            throw new ConflictException("이미 등록된 테마 이름입니다. 다른 이름을 입력해주세요.");
+        }
         Theme theme = new Theme(name, description, thumbnail);
 
         return themeRepository.save(theme);
     }
 
-    private void validateNotDuplicated(String name) {
-        if (themeRepository.existsByName(name)) {
-            throw new InvalidRequestException("이미 존재하는 테마 이름입니다.");
-        }
-    }
-
     @Transactional(readOnly = true)
-    public List<Theme> list(){
+    public List<Theme> list() {
         return themeRepository.findAll();
     }
 
     @Transactional(readOnly = true)
-    public List<Theme> findPopularThemes(int days, LocalDate now, int size){
+    public List<Theme> findPopularThemes(int days, LocalDate now, int size) {
         LocalDate startDate = now.minusDays(days);
         LocalDate endDate = now.minusDays(1);
 
@@ -46,6 +45,10 @@ public class ThemeService {
 
     @Transactional
     public void delete(Long id) {
+        if (reservationRepository.existsByThemeId(id)) {
+            throw new ConflictException("예약이 존재하는 테마는 삭제할 수 없습니다. 먼저 해당 예약들을 삭제해주세요.");
+        }
+
         themeRepository.deleteById(id);
     }
 }
